+++ demo/domains/legal/entity_resolvers.py
@@ -1,9 +1,18 @@
+from typing import Optional
 import logging
+
+from rich.console import Console
+from neo4j import GraphDatabase
+
+from proscenium.scripts.entity_resolver import load_entity_resolver
+from proscenium.scripts.entity_resolver import vector_db
 from proscenium.scripts.entity_resolver import Resolver
 
-log = logging.getLogger(__name__)
+from proscenium.core import Prop
 
 default_embedding_model_id = "all-MiniLM-L6-v2"
+
+log = logging.getLogger(__name__)
 
 case_resolver = Resolver(
     "MATCH (cr:CaseRef) RETURN cr.text AS text",
@@ -17,4 +26,57 @@ judge_resolver = Resolver(
     "resolve_judgerefs",
 )
 
-resolvers = [case_resolver, judge_resolver]
+
+class EntityResolvers(Prop):
+
+    def __init__(
+        self,
+        milvus_uri: str,
+        embedding_model_id: str,
+        neo4j_uri: str,
+        neo4j_username: str,
+        neo4j_password: str,
+        console: Optional[Console] = None,
+    ) -> None:
+        super().__init__(console)
+        self.milvus_uri = milvus_uri
+        self.embedding_model_id = embedding_model_id
+        self.neo4j_uri = neo4j_uri
+        self.neo4j_username = neo4j_username
+        self.neo4j_password = neo4j_password
+
+        self.resolvers = [case_resolver, judge_resolver]
+
+    def already_built(self) -> bool:
+        client = vector_db(self.milvus_uri, overwrite=False)
+        collections = client.list_collections()
+        try:
+            for resolver in self.resolvers:
+                collection_name = resolver.collection_name
+                if collection_name not in collections:
+                    return False
+                # row_count = client.get_collection_stats(collection_name)["row_count"]
+        finally:
+            client.close()
+
+        return True
+
+    def build(self, force: bool = False):
+
+        if force or self.already_built():
+            log.info("Entity resolver already built.")
+            return
+
+        driver = GraphDatabase.driver(
+            self.neo4j_uri, auth=(self.neo4j_username, self.neo4j_password)
+        )
+
+        load_entity_resolver(
+            driver,
+            self.resolvers,
+            self.embedding_model_id,
+            self.milvus_uri,
+            console=self.console,
+        )
+
+        driver.close()
