+++ demo/domains/literature/query_handler.py
@@ -1,11 +1,12 @@
 from typing import Generator
-from typing import Callable
 
 import logging
 
 from proscenium.verbs.vector_database import embedding_function
 from proscenium.verbs.vector_database import vector_db
 from proscenium.scripts.rag import answer_question
+
+from proscenium.core import Character
 
 from demo.config import default_model_id
 
@@ -21,29 +22,34 @@ user_prompt = f"What is your question about {', '.join([b.title for b in books])
 default_question = "What did Hermes say to Prometheus about giving fire to humans?"
 
 
-def make_handler(
-    generator_model_id: str,
-    milvus_uri: str,
-    embedding_model_id: str,
-    admin_channel_id: str,
-) -> Callable[[tuple[str, str, str]], Generator[tuple[str, str], None, None]]:
+class LiteratureExpert(Character):
 
-    vector_db_client = vector_db(milvus_uri)
-    log.info("Vector db at uri %s", milvus_uri)
+    def __init__(
+        self,
+        generator_model_id: str,
+        milvus_uri: str,
+        embedding_model_id: str,
+        admin_channel_id: str,
+    ):
+        super().__init__(admin_channel_id=admin_channel_id)
+        self.generator_model_id = generator_model_id
+        self.milvus_uri = milvus_uri
+        self.embedding_model_id = embedding_model_id
 
-    embedding_fn = embedding_function(embedding_model_id)
-    log.info("Embedding model %s", embedding_model_id)
+        self.vector_db_client = vector_db(self.milvus_uri)
+        log.info("Vector db at uri %s", milvus_uri)
+
+        self.embedding_fn = embedding_function(embedding_model_id)
+        log.info("Embedding model %s", embedding_model_id)
 
     def handle(
-        channel_id: str, speaker_id: str, question: str
+        self, channel_id: str, speaker_id: str, utterance: str
     ) -> Generator[tuple[str, str], None, None]:
 
         yield channel_id, answer_question(
-            question,
-            generator_model_id,
-            vector_db_client,
-            embedding_fn,
+            utterance,
+            self.generator_model_id,
+            self.vector_db_client,
+            self.embedding_fn,
             default_collection_name,
         )
-
-    return handle
